pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avcp-template"
version = "0.1.0"
description = "AVCP template repository (prompts + docs + code-enforced tooling)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "pyarrow>=12.0",
  "pyyaml>=6.0",
  "tqdm>=4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "pre-commit>=3.0.0",
  "types-PyYAML>=6.0.12",
  "pandas-stubs>=2.0.0",
  "types-setuptools>=69.0.0",
  "jinja2>=3.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
